@@ -73,20 +73,19 @@
 		<asciidoctorj-diagram.version>2.1.0</asciidoctorj-diagram.version>
 
 		<byte-buddy.version>1.11.0</byte-buddy.version>
-		<cdi>2.0.SP1</cdi>
 		<checkstyle.version>8.40</checkstyle.version>
 		<cypher-dsl.version>2021.4.0</cypher-dsl.version>
 		<dist.id>spring-data-neo4j</dist.id>
 		<dist.key>SDNEO4J</dist.key>
 		<flatten-maven-plugin.version>1.2.5</flatten-maven-plugin.version>
-		<jacoco-maven-plugin.version>0.8.6</jacoco-maven-plugin.version>
+		<jacoco-maven-plugin.version>0.8.7</jacoco-maven-plugin.version>
 		<java-module-name>spring.data.neo4j</java-module-name>
-		<java.version>1.8</java.version>
+		<java.version>17</java.version>
 		<jaxb.version>2.3.1</jaxb.version>
-		<jqassistant-dashboard-plugin.version>1.9.0</jqassistant-dashboard-plugin.version>
+		<jqassistant-dashboard-plugin.version>1.10.0</jqassistant-dashboard-plugin.version>
 		<jqassistant.plugin.git.version>1.8.0</jqassistant.plugin.git.version>
-		<jqassistant.plugin.version>1.9.1</jqassistant.plugin.version>
-		<jqassistant.version>1.9.1</jqassistant.version>
+		<jqassistant.plugin.version>1.10.1</jqassistant.plugin.version>
+		<jqassistant.version>1.10.1</jqassistant.version>
 		<junit-cc-testcontainer>2021.0.1</junit-cc-testcontainer>
 		<maven-checkstyle-plugin.version>3.1.2</maven-checkstyle-plugin.version>
 		<maven-deploy-plugin.version>3.0.0-M1</maven-deploy-plugin.version>
